FEAT: broke up connection code
