+++ server/src/main.rs
@@ -1,3 +1,5 @@
+mod connection_manager;
+
 use actix_session::{storage::CookieSessionStore, Session, SessionMiddleware};
 use actix_web::{
     cookie::Key,
@@ -16,25 +18,12 @@ use tokio::sync::{
 };
 use webrtc::{
     api::{media_engine::MediaEngine, APIBuilder},
-    data_channel::{data_channel_message::DataChannelMessage, RTCDataChannel},
-    ice_transport::{
-        ice_candidate::{RTCIceCandidate, RTCIceCandidateInit},
-        ice_gatherer_state::RTCIceGathererState,
-        ice_server::RTCIceServer,
-    },
+    ice_transport::ice_candidate::RTCIceCandidateInit,
     interceptor::registry::Registry,
-    peer_connection::{
-        configuration::RTCConfiguration,
-        peer_connection_state::RTCPeerConnectionState,
-        policy::{
-            bundle_policy::RTCBundlePolicy, ice_transport_policy::RTCIceTransportPolicy,
-            rtcp_mux_policy::RTCRtcpMuxPolicy,
-        },
-        sdp::session_description::RTCSessionDescription,
-        signaling_state::RTCSignalingState,
-        RTCPeerConnection,
-    },
+    peer_connection::RTCPeerConnection,
 };
+
+use crate::connection_manager::ConnectionMessage;
 
 #[derive(Deserialize, Serialize)]
 struct OfferDescription {
@@ -51,28 +40,6 @@ struct IceCandidates {
     candidates: Vec<RTCIceCandidateInit>,
 }
 
-enum ConnectionMessage {
-    NewConnection {
-        offer: String,
-        tx: Sender<ConnectionMessage>,
-        resp: oneshot::Sender<Option<(String, u32)>>,
-    },
-    AddRemoteCandidate {
-        id: u32,
-        candidate: RTCIceCandidateInit,
-        resp: oneshot::Sender<()>,
-    },
-    AddLocalCandidate {
-        id: u32,
-        candidate: RTCIceCandidateInit,
-        resp: oneshot::Sender<()>,
-    },
-    GetIceCandidates {
-        id: u32,
-        resp: oneshot::Sender<Option<IceCandidates>>,
-    },
-    Cleanup,
-}
 
 // post new_offer
 // returns answer string
@@ -155,7 +122,7 @@ async fn ice_candidate(
 }
 
 #[get("/ice_candidate")]
-async fn get_ice_candidates(session: Session, data: Data<Sender<ConnectionMessage>>) -> Option<Json<IceCandidates>> {
+async fn get_ice_candidates(session: Session, data: Data<Sender<ConnectionMessage>>) -> Option<Json<Vec<RTCIceCandidateInit>>> {
     let (resp_tx, resp_rx) = oneshot::channel();
 
     let id = session.get("id").ok().flatten()?;
@@ -223,120 +190,19 @@ async fn main() {
     while let Some(message) = rx.recv().await {
         match message {
             ConnectionMessage::NewConnection { offer, tx, resp } => {
-                let config = RTCConfiguration {
-                    ice_servers: vec![RTCIceServer {
-                        urls: vec!["stun:stun.l.google.com:19302".to_owned()],
-                        ..Default::default()
-                    }],
-                    ice_transport_policy: RTCIceTransportPolicy::All,
-                    bundle_policy: RTCBundlePolicy::Balanced,
-                    rtcp_mux_policy: RTCRtcpMuxPolicy::Require,
-                    ..Default::default()
-                };
 
-                let pc = api.new_peer_connection(config).await.ok();
+                let connection = connection_manager::new_connection(&api, tx.clone(), offer).await;
 
-                if pc.is_none() {
+                if connection.is_none() {
                     let _ = resp.send(None);
                     continue;
                 }
 
-                let pc = Arc::new(pc.unwrap());
-
-                let id = 11; // TODO: new id
-                
-                pc.on_ice_candidate(Box::new(move |candidate: Option<RTCIceCandidate>| {
-                    let tx2 = tx.clone();
-
-                    Box::pin(async move {
-                        if let Some(candidate) = candidate {
-                            if let Ok(candidate_json) = candidate.to_json() {
-                                let (resp_tx, resp_rx) = oneshot::channel();
-
-                                let _ = tx2.send(ConnectionMessage::AddLocalCandidate {
-                                    id,
-                                    candidate: candidate_json,
-                                    resp: resp_tx,
-                                }).await;
-
-                                let _ = resp_rx.await;
-                            }
-                        }
-                    })
-                }));
-
-                pc.on_signaling_state_change(Box::new(move |s: RTCSignalingState| {
-                    println!("Signaling State has changed: {s}");
-                    Box::pin(async {})
-                }));
-
-                pc.on_ice_gathering_state_change(Box::new(move |s: RTCIceGathererState| {
-                    println!("Ice Gathering State has changed: {s}");
-
-                    Box::pin(async {})
-                }));
-
-                pc.on_peer_connection_state_change(Box::new(move |s: RTCPeerConnectionState| {
-                    println!("Peer Connection State has changed: {s}");
-
-                    if s == RTCPeerConnectionState::Failed {
-                        println!("Peer Connection has gone to failed exiting");
-                    }
-
-                    Box::pin(async {})
-                }));
-
-                pc.on_data_channel(Box::new(move |d: Arc<RTCDataChannel>| {
-                    let d_label = d.label().to_owned();
-                    let d_id = d.id();
-                    println!("New data channel {d_label} {d_id}");
-
-                    Box::pin(async move {
-                        let d_label2 = d_label.clone();
-
-                        d.on_close(Box::new(move || {
-                            println!("Data channel closed");
-                            Box::pin(async {})
-                        }));
-
-                        d.on_open(Box::new(move || {
-                            println!("Data channel {d_label} {d_id} opened");
-                            Box::pin(async {})
-                        }));
-
-                        d.on_message(Box::new(move |msg: DataChannelMessage| {
-                            let msg_str = String::from_utf8(msg.data.to_vec()).unwrap();
-                            println!("Message from data channel '{d_label2}': '{msg_str}'");
-                            Box::pin(async {})
-                        }));
-                    })
-                }));
-
-                let session_desc = RTCSessionDescription::offer(offer.clone()).ok();
-
-                if session_desc.is_none() {
-                    let _ = resp.send(None);
-                    continue;
-                }
-
-                let session_desc = session_desc.unwrap();
-
-                let _ = pc.set_remote_description(session_desc).await.ok();
-
-                let answer = pc.create_answer(None).await.ok();
-
-                if answer.is_none() {
-                    let _ = resp.send(None);
-                    continue;
-                }
-
-                let answer = answer.unwrap();
-
-                let _ = pc.set_local_description(answer.clone()).await; // dammit
+                let (pc, id, answer) = connection.unwrap();
 
                 connections.insert(id, pc);
 
-                let _ = resp.send(Some((answer.sdp, id)));
+                let _ = resp.send(Some((answer, id)));
             }
             ConnectionMessage::AddRemoteCandidate {
                 id,
@@ -365,9 +231,7 @@ async fn main() {
                 let pc_candidates = candidates.remove(&id);
 
                 println!("{:?}", pc_candidates);
-                let _ = resp.send(pc_candidates.map(|c| IceCandidates {
-                    candidates: c
-                }));
+                let _ = resp.send(pc_candidates);
             }
             ConnectionMessage::Cleanup => {
                 for (_, value) in connections.iter() {
